services/text_analysis/requirements.txt

fastapi
openai
google-cloud-language
ibm-watson
cohere
requests
